Refactor libraries in CurrencyConverter and Budget class to resolve errors
